Consolidate docgen example collision regression coverage
